+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - New configuration section: python.requirements
 * Fix potential memory leak due to gradle Project objects caching (#22)
 * Fix concurrent virtualenv installation in multi-module project with parallel execution enabled (#19)
+* Fix concurrent pipInstall into the same environment with parallel execution
 
 ### 2.3.0 (2021-03-01)
 * Support python installation from Windows Store (#14)